Move ReleaseTransactionBuilderBuildAmountToTest and ReleaseTransactionBuilderBuildEmptyWalletToTest into ReleaseTransactionBuilderTest as @Nested BuildAmountToTest and BuildEmptyWalletToTest. Keep all existing mock/regtest coverage under @Nested MockWalletReleaseBuilderTest and remove the two standalone test classes.
